fix workflow error and sort import
